refactor: Refactor user id foreign keys to type 'long' instead of 'int'

+++ DeepAzureServer/Models/Entities/Conversation.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public DateTime LastMessageAt { get; set; }
-        public int User1Id { get; set; }
-        public int User2Id { get; set; }
+        public long User1Id { get; set; }
+        public long User2Id { get; set; }
 
         public User? User1 { get; set; }
         public User? User2 { get; set; }
